Refactor event and task management system
- Updated Invitation class to use Long for ID and added JPA annotations.
- Modified InvitationManager to handle Long ID type.
- Refactored Task class with JPA annotations.
- Enhanced People class with JPA annotations and additional fields.
- Removed obsolete TaskController and mainPage.html files.
- Improved script.js for dynamic loading of tasks, teachers, and calendar events.
- Added EventRepository and EventService for event management.
- Created PeopleRepository and PeopleService for user management.
- Implemented ApiController to handle API requests for tasks, events, and teachers.
- Updated styles.css for improved UI and added event styling.
- Created index.html as the main entry point for the application.
-Events is synchronized with DB too!!!

+++ demo/src/main/java/com/example/demo/javaSrc/eventsANDtask/InvitationManager.java
@@ -1,7 +1,13 @@
 package com.example.demo.javaSrc.eventsANDtask;
 
-import java.sql.*;
-import java.util.*;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
 
 public class InvitationManager {
     private static final String URL = "jdbc:mysql://localhost:3306/PeopleAndEvents";
@@ -67,7 +73,7 @@ public class InvitationManager {
             try (ResultSet rs = stmt.executeQuery()) {
                 while (rs.next()) {
                     Invitation invitation = new Invitation(
-                        rs.getInt("id"),
+                        rs.getLong("id"),
                         rs.getInt("event_id"),
                         rs.getInt("user_id"),
                         rs.getTimestamp("created_at"),
